算法重写

+++ App/Inc/motor.h
@@ -5,9 +5,9 @@ extern float32_t filter_wight[3];
 extern uint16 motor_pulse;
 
 extern float32_t motor_param[3];
-extern float64_t expected_motor_out;
-extern float64_t motor_errors[3];
-extern float64_t motor_out;
+extern float32_t expected_motor_out;
+extern float32_t motor_errors[3];
+extern float32_t motor_out;
 
 void encoder();
 void motor();
